Update asset_path in GitHub action for soccr plugin
The GitHub action for the soccr WordPress plugin has been updated to change the asset_path value. Instead of using a path within a release-artifact folder, the version number is now being set directly within the soccr-reloaded.zip file. This modification enhances the accuracy of versioning and improves the overall efficiency of the build process.